+++ src/plugins/core/cell.ts
@@ -4,15 +4,23 @@ import {
   groupItemIdsByZones,
   iterateItemIdsPositions,
 } from "../../helpers/data_normalization";
-import { deepEquals, isObjectEmptyRecursive, range, replaceNewLines } from "../../helpers/misc";
+import {
+  deepCopy,
+  deepEquals,
+  isObjectEmptyRecursive,
+  range,
+  replaceNewLines,
+} from "../../helpers/misc";
 
+import { DEFAULT_STYLE } from "../../constants";
 import { toXC } from "../../helpers/coordinates";
 import { CorePlugin } from "../core_plugin";
+import { DefaultPlugin, defaultStyle, defaultValue } from "./default";
 import { SettingsPlugin } from "./settings";
 import { SheetPlugin } from "./sheet";
 
 import { getDateTimeFormat } from "../../helpers/locale";
-import { isInside } from "../../helpers/zones";
+import { cellPositions, getZoneArea, isInside } from "../../helpers/zones";
 import { Cell } from "../../types/cells";
 import {
   AddColumnsRowsCommand,
@@ -20,6 +28,7 @@ import {
   CommandResult,
   CoreCommand,
   PositionDependentCommand,
+  SetFormattingCommand,
   UpdateCellCommand,
 } from "../../types/commands";
 import { CellPosition, HeaderIndex, RangeAdapterFunctions, UID } from "../../types/misc";
@@ -53,7 +62,7 @@ interface CoreState {
  * cell and sheet content.
  */
 export class CellPlugin extends CorePlugin<typeof CellPlugin, CoreState> implements CoreState {
-  static readonly dependencies = [SheetPlugin, SettingsPlugin] as const;
+  static readonly dependencies = [SheetPlugin, SettingsPlugin, DefaultPlugin] as const;
   static getters = [
     "getCells",
     "getTranslatedCellFormula",
@@ -61,10 +70,21 @@ export class CellPlugin extends CorePlugin<typeof CellPlugin, CoreState> impleme
     "getFormulaString",
     "getFormulaMovedInSheet",
     "getCell",
+    "getCellStyle",
+    "getCellFormat",
   ] as const;
   readonly nextId = 1;
   public readonly cells: { [sheetId: string]: { [id: string]: Cell } } = {};
   previousLocale: Locale = this.getters.getLocale();
+
+  /**
+   * Snapshots of the default style/format taken in `beforeHandle`, before the
+   * DefaultPlugin overwrites them in its `handle`. The cell-side of a formatting
+   * command needs the *previous* defaults to know which values must be baked
+   * into cells to preserve their appearance.
+   */
+  private previousDefaultStyle: defaultStyle | undefined;
+  private previousDefaultFormat: defaultValue<Format> | undefined;
 
   adaptRanges(adapters: RangeAdapterFunctions) {
     for (const sheet of Object.keys(this.cells)) {
@@ -99,8 +119,23 @@ export class CellPlugin extends CorePlugin<typeof CellPlugin, CoreState> impleme
         return !cmd.cellId || this.cells[cmd.sheetId]?.[cmd.cellId]
           ? CommandResult.Success
           : CommandResult.InvalidCellId;
+      case "SET_FORMATTING":
+        return this.checkUselessSetFormatting(cmd);
       default:
         return CommandResult.Success;
+    }
+  }
+
+  beforeHandle(cmd: CoreCommand) {
+    switch (cmd.type) {
+      case "SET_FORMATTING":
+      case "CLEAR_FORMATTING":
+        // Snapshot the defaults before the DefaultPlugin overwrites them in its
+        // `handle`, so the cell baking (done in our `handle`) can preserve the
+        // previous appearance of the impacted cells.
+        this.previousDefaultStyle = deepCopy(this.getters.getSheetDefaultStyle(cmd.sheetId));
+        this.previousDefaultFormat = deepCopy(this.getters.getSheetDefaultFormat(cmd.sheetId));
+        break;
     }
   }
 
@@ -112,6 +147,18 @@ export class CellPlugin extends CorePlugin<typeof CellPlugin, CoreState> impleme
         } else {
           this.handleAddColumnsRows(cmd, this.copyRowStyle.bind(this));
         }
+        break;
+      case "SET_FORMATTING":
+        if (cmd.style !== undefined) {
+          this.setStyle(cmd.sheetId, cmd.target, cmd.style);
+        }
+        if (cmd.format !== undefined) {
+          this.setFormat(cmd.sheetId, cmd.target, cmd.format);
+        }
+        break;
+      case "CLEAR_FORMATTING":
+        this.setStyle(cmd.sheetId, cmd.target, DEFAULT_STYLE);
+        this.setFormat(cmd.sheetId, cmd.target, null);
         break;
       case "UPDATE_CELL":
         this.updateCell(cmd.sheetId, cmd.col, cmd.row, cmd);
@@ -577,6 +624,398 @@ export class CellPlugin extends CorePlugin<typeof CellPlugin, CoreState> impleme
     return this.getters.getCellById(cellId);
   }
 
+  getCellStyle(position: CellPosition): Style {
+    const cell = this.getters.getCell(position);
+    return { ...this.getters.getCellDefaultStyle(position), ...cell?.style };
+  }
+
+  getCellFormat(position: CellPosition): Format | undefined {
+    const cell = this.getters.getCell(position);
+    if (cell?.format !== undefined) {
+      return cell?.format;
+    }
+    return this.getters.getCellDefaultFormat(position);
+  }
+
+  // ---------------------------------------------------------------------------
+  // Formatting (cell side of SET_FORMATTING / CLEAR_FORMATTING)
+  //
+  // The default style/format state is owned and updated by the DefaultPlugin.
+  // Here we only update the individual cells: clearing cell styles overridden by
+  // a new default, and baking the previous defaults into cells whose appearance
+  // would otherwise change. These run in `handle`, after the DefaultPlugin has
+  // updated its state, so the default comparisons (in `updateCell`) use the new
+  // defaults; the *previous* defaults are read from the snapshots taken in
+  // `beforeHandle`.
+  // ---------------------------------------------------------------------------
+
+  private setFormat(sheetId: UID, zones: Zone[], format: Format | null) {
+    zones = recomputeZones(zones);
+    const { numberOfCols, numberOfRows } = this.getters.getSheetSize(sheetId);
+    const sheetArea = numberOfCols * numberOfRows;
+    for (const zone of zones) {
+      const defaultCol = zone.bottom - zone.top + 1 > numberOfRows / 2;
+      const defaultRow = zone.right - zone.left + 1 > numberOfCols / 2;
+      if (defaultRow && defaultCol && getZoneArea(zone) > sheetArea / 2) {
+        this.setSheetFormat(sheetId, zone, format);
+      } else if (defaultCol) {
+        this.setColsFormat(sheetId, zone, format ?? "");
+      } else if (defaultRow) {
+        this.setRowsFormat(sheetId, zone, format ?? "");
+      } else {
+        this.updateCellsFormat(sheetId, zone, format ?? "");
+      }
+    }
+  }
+
+  private setSheetFormat(sheetId: UID, zone: Zone, format: Format | null) {
+    this.updateCellsFormat(sheetId, zone, null);
+    const sheetZone = this.getters.getSheetZone(sheetId);
+    const horizontalZone = this.getters.getRowsZone(sheetId, zone.top, zone.bottom);
+    const externalHorizontalZones = recomputeZones([horizontalZone], [zone]);
+    const defaults = this.getDefaultFormatInCell(sheetId, externalHorizontalZones, {
+      shouldUseDefaultSheet: true,
+      shouldUseDefaultRow: true,
+    });
+    const verticalZone = this.getters.getColsZone(sheetId, zone.left, zone.right);
+    const externalVerticalZones = recomputeZones([verticalZone], [zone]);
+    defaults.push(
+      ...this.getDefaultFormatInCell(sheetId, externalVerticalZones, {
+        shouldUseDefaultSheet: true,
+        shouldUseDefaultCol: true,
+      })
+    );
+    const externalCornerZones = recomputeZones([sheetZone], [horizontalZone, verticalZone]);
+    defaults.push(
+      ...this.getDefaultFormatInCell(sheetId, externalCornerZones, { shouldUseDefaultSheet: true })
+    );
+    for (const [position, value] of defaults) {
+      this.updateCellFormat(position, value);
+    }
+  }
+
+  private setColsFormat(sheetId: UID, zone: Zone, format: Format) {
+    this.updateCellsFormat(sheetId, zone, null);
+    const leftoverZones = recomputeZones(
+      [this.getters.getColsZone(sheetId, zone.left, zone.right)],
+      [zone]
+    );
+    const defaults = this.getDefaultFormatInCell(sheetId, leftoverZones, {
+      shouldUseDefaultSheet: true,
+      shouldUseDefaultCol: true,
+    });
+    const rowOverlap = Object.keys(this.previousDefaultFormat?.rowDefault ?? {});
+    for (let col = zone.left; col <= zone.right; col++) {
+      for (const rowIndex of rowOverlap) {
+        const row = parseInt(rowIndex);
+        if (zone.top <= row && row <= zone.bottom) {
+          this.updateCellFormat({ col, row, sheetId }, format);
+        }
+      }
+    }
+    for (const [position, value] of defaults) {
+      this.updateCellFormat(position, value);
+    }
+  }
+
+  private setRowsFormat(sheetId: UID, zone: Zone, format: Format) {
+    this.updateCellsFormat(sheetId, zone, null);
+    const leftoverZones = recomputeZones(
+      [this.getters.getRowsZone(sheetId, zone.top, zone.bottom)],
+      [zone]
+    );
+    const defaults = this.getDefaultFormatInCell(sheetId, leftoverZones, {
+      shouldUseDefaultSheet: true,
+      shouldUseDefaultCol: true,
+      shouldUseDefaultRow: true,
+    });
+    for (const [position, value] of defaults) {
+      this.updateCellFormat(position, value);
+    }
+  }
+
+  private updateCellsFormat(sheetId: UID, zone: Zone, format: Format | null) {
+    for (let col = zone.left; col <= zone.right; col++) {
+      for (let row = zone.top; row <= zone.bottom; row++) {
+        this.updateCellFormat({ sheetId, col, row }, format);
+      }
+    }
+  }
+
+  private updateCellFormat(position: CellPosition, format: Format | null) {
+    if ((format ?? "") !== (this.getters.getCellDefaultFormat(position) ?? "")) {
+      this.dispatch("UPDATE_CELL", {
+        sheetId: position.sheetId,
+        col: position.col,
+        row: position.row,
+        format,
+      });
+    } else {
+      this.dispatch("UPDATE_CELL", {
+        sheetId: position.sheetId,
+        col: position.col,
+        row: position.row,
+        format: null,
+      });
+    }
+  }
+
+  private getDefaultFormatInCell(
+    sheetId: UID,
+    zones: Zone[],
+    priorities: {
+      shouldUseDefaultCol?: boolean;
+      shouldUseDefaultRow?: boolean;
+      shouldUseDefaultSheet?: boolean;
+    }
+  ): [CellPosition, Format][] {
+    const defaults: [CellPosition, Format][] = [];
+    const formatSheet = this.previousDefaultFormat;
+    for (const position of zones.flatMap((zone) => cellPositions(sheetId, zone))) {
+      const cellFormat = this.getters.getCell(position)?.format;
+      if (cellFormat) {
+        continue;
+      }
+      const rowDefault = formatSheet?.rowDefault?.[position.row];
+      if (rowDefault) {
+        if (priorities.shouldUseDefaultRow) {
+          defaults.push([position, rowDefault]);
+        }
+        continue;
+      }
+      const colDefault = formatSheet?.colDefault?.[position.col];
+      if (colDefault) {
+        if (priorities.shouldUseDefaultCol) {
+          defaults.push([position, colDefault]);
+        }
+        continue;
+      }
+      const sheetDefault = formatSheet?.sheetDefault;
+      if (sheetDefault) {
+        if (priorities.shouldUseDefaultSheet) {
+          defaults.push([position, sheetDefault]);
+        }
+        continue;
+      }
+    }
+    return defaults;
+  }
+
+  private setStyle(sheetId: UID, zones: Zone[], style: Style) {
+    zones = recomputeZones(zones);
+    const { numberOfCols, numberOfRows } = this.getters.getSheetSize(sheetId);
+    const sheetArea = numberOfCols * numberOfRows;
+    for (const zone of zones) {
+      const defaultCol = zone.bottom - zone.top + 1 > numberOfRows / 2;
+      const defaultRow = zone.right - zone.left + 1 > numberOfCols / 2;
+      if (defaultRow && defaultCol && getZoneArea(zone) > sheetArea / 2) {
+        this.setSheetStyle(sheetId, zone, style);
+      } else if (defaultCol) {
+        this.setColsStyle(sheetId, zone, style);
+      } else if (defaultRow) {
+        this.setRowsStyle(sheetId, zone, style);
+      } else {
+        this.updateCellsStyle(sheetId, zone, style);
+      }
+    }
+  }
+
+  private setSheetStyle(sheetId: UID, zone: Zone, style: Style) {
+    this.clearCellStyle(sheetId, zone, style);
+    const sheetZone = this.getters.getSheetZone(sheetId);
+    const horizontalZone = this.getters.getRowsZone(sheetId, zone.top, zone.bottom);
+    const externalHorizontalZones = recomputeZones([horizontalZone], [zone]);
+    const defaults = this.getPartialDefaultStyleInCell(sheetId, externalHorizontalZones, style, {
+      shouldUseDefaultSheet: true,
+      shouldUseDefaultRow: true,
+    });
+    const verticalZone = this.getters.getColsZone(sheetId, zone.left, zone.right);
+    const externalVerticalZones = recomputeZones([verticalZone], [zone]);
+    defaults.push(
+      ...this.getPartialDefaultStyleInCell(sheetId, externalVerticalZones, style, {
+        shouldUseDefaultSheet: true,
+        shouldUseDefaultCol: true,
+      })
+    );
+    const externalCornerZones = recomputeZones([sheetZone], [horizontalZone, verticalZone]);
+    defaults.push(
+      ...this.getPartialDefaultStyleInCell(sheetId, externalCornerZones, style, {
+        shouldUseDefaultSheet: true,
+      })
+    );
+    for (const [position, value] of defaults) {
+      this.updateCellStyle(position, value);
+    }
+  }
+
+  private setColsStyle(sheetId: UID, zone: Zone, style: Style) {
+    this.clearCellStyle(sheetId, zone, style);
+    const leftoverZones = recomputeZones(
+      [this.getters.getColsZone(sheetId, zone.left, zone.right)],
+      [zone]
+    );
+    const defaults = this.getPartialDefaultStyleInCell(sheetId, leftoverZones, style, {
+      shouldUseDefaultSheet: true,
+      shouldUseDefaultCol: true,
+    });
+    const overlapUpdate = new PositionMap<Style>();
+    const styleSheet = this.previousDefaultStyle;
+    for (const key in style) {
+      const rowOverlap = Object.keys(styleSheet?.[key]?.rowDefault ?? {});
+      for (let col = zone.left; col <= zone.right; col++) {
+        for (const rowIndex of rowOverlap) {
+          const row = parseInt(rowIndex);
+          if (zone.top <= row && row <= zone.bottom) {
+            const position = { col, row, sheetId };
+            const s = overlapUpdate.get(position);
+            if (s) {
+              s[key] = style[key];
+            } else {
+              const newStyle = {};
+              newStyle[key] = style[key];
+              overlapUpdate.set(position, newStyle);
+            }
+          }
+        }
+      }
+    }
+    for (const [position, overlapStyle] of overlapUpdate.entries()) {
+      this.updateCellStyle(position, overlapStyle);
+    }
+    for (const [position, value] of defaults) {
+      this.updateCellStyle(position, value);
+    }
+  }
+
+  private setRowsStyle(sheetId: UID, zone: Zone, style: Style) {
+    this.clearCellStyle(sheetId, zone, style);
+    const leftoverZones = recomputeZones(
+      [this.getters.getRowsZone(sheetId, zone.top, zone.bottom)],
+      [zone]
+    );
+    const defaults = this.getPartialDefaultStyleInCell(sheetId, leftoverZones, style, {
+      shouldUseDefaultSheet: true,
+      shouldUseDefaultCol: true,
+      shouldUseDefaultRow: true,
+    });
+    for (const [position, value] of defaults) {
+      this.updateCellStyle(position, value);
+    }
+  }
+
+  private updateCellsStyle(sheetId: UID, zone: Zone, style: Style) {
+    for (let col = zone.left; col <= zone.right; col++) {
+      for (let row = zone.top; row <= zone.bottom; row++) {
+        this.updateCellStyle({ sheetId, col, row }, style);
+      }
+    }
+  }
+
+  private updateCellStyle(position: CellPosition, style: Style) {
+    const cell = this.getters.getCell(position);
+    const cellStyle = { ...cell?.style, ...style };
+    this.dispatch("UPDATE_CELL", {
+      sheetId: position.sheetId,
+      col: position.col,
+      row: position.row,
+      style: cellStyle,
+    });
+  }
+
+  private clearCellStyle(sheetId: UID, zone: Zone, style: Style) {
+    for (let row = zone.top; row <= zone.bottom; row++) {
+      for (const cellId of this.getters.getRowCellIds(sheetId, row)) {
+        const col = this.getters.getCellPosition(cellId).col;
+        if (col < zone.left || zone.right < col) {
+          continue;
+        }
+        let cellStyle = this.getters.getCellById(cellId)?.style;
+        if (!cellStyle) {
+          continue;
+        }
+        cellStyle = { ...cellStyle };
+        let dispatch = false;
+        for (const key in style) {
+          if (cellStyle[key] !== undefined) {
+            dispatch = true;
+            delete cellStyle[key];
+          }
+        }
+        if (dispatch) {
+          this.dispatch("UPDATE_CELL", {
+            sheetId,
+            col,
+            row,
+            style: Object.keys(cellStyle).length === 0 ? null : cellStyle,
+          });
+        }
+      }
+    }
+  }
+
+  private getPartialDefaultStyleInCell(
+    sheetId: UID,
+    zones: Zone[],
+    newDefaultStyle: Style,
+    priorities: {
+      shouldUseDefaultCol?: boolean;
+      shouldUseDefaultRow?: boolean;
+      shouldUseDefaultSheet?: boolean;
+    }
+  ): [CellPosition, Style][] {
+    const partialDefaults: [CellPosition, Style][] = [];
+    const styleSheet = this.previousDefaultStyle;
+    if (!styleSheet) {
+      return partialDefaults;
+    }
+    for (const position of zones.flatMap((zone) => cellPositions(sheetId, zone))) {
+      const cellStyle = this.getters.getCell(position)?.style ?? {};
+      const deltaStyle: Style = {};
+      let hasDelta = false;
+      for (const key in newDefaultStyle) {
+        if (key in cellStyle) {
+          continue;
+        }
+        const defaults = styleSheet[key];
+        if (!defaults) {
+          continue;
+        }
+        const rowDefault = defaults.rowDefault?.[position.row];
+        if (rowDefault !== undefined) {
+          if (priorities.shouldUseDefaultRow) {
+            deltaStyle[key] = rowDefault;
+            hasDelta = true;
+          }
+          continue;
+        }
+        const colDefault = defaults.colDefault?.[position.col];
+        if (colDefault !== undefined) {
+          if (priorities.shouldUseDefaultCol) {
+            deltaStyle[key] = colDefault;
+            hasDelta = true;
+          }
+          continue;
+        }
+        const sheetDefault = defaults.sheetDefault;
+        if (sheetDefault !== undefined) {
+          if (priorities.shouldUseDefaultSheet) {
+            deltaStyle[key] = sheetDefault;
+            hasDelta = true;
+          }
+          continue;
+        }
+        if (newDefaultStyle[key] !== DEFAULT_STYLE[key]) {
+          deltaStyle[key] = DEFAULT_STYLE[key];
+          hasDelta = true;
+        }
+      }
+      if (hasDelta) {
+        partialDefaults.push([position, deltaStyle]);
+      }
+    }
+    return partialDefaults;
+  }
+
   private changeCellsDateFormatWithLocale(oldLocale: Locale, newLocale: Locale) {
     for (const sheetId of this.getters.getSheetIds()) {
       for (const cell of this.getters.getCells(sheetId)) {
@@ -601,6 +1040,29 @@ export class CellPlugin extends CorePlugin<typeof CellPlugin, CoreState> impleme
         }
       }
     }
+  }
+
+  private checkUselessSetFormatting(cmd: SetFormattingCommand): CommandResult {
+    const { sheetId, target } = cmd;
+    const hasStyle = "style" in cmd;
+    const hasFormat = "format" in cmd;
+    if (!hasStyle && !hasFormat) {
+      return CommandResult.NoChanges;
+    }
+    for (const zone of recomputeZones(target)) {
+      for (let col = zone.left; col <= zone.right; col++) {
+        for (let row = zone.top; row <= zone.bottom; row++) {
+          const position = { sheetId, col, row };
+          if (
+            (hasStyle && !deepEquals(this.getCellStyle(position), cmd.style)) ||
+            (hasFormat && this.getCellFormat(position) !== cmd.format)
+          ) {
+            return CommandResult.Success;
+          }
+        }
+      }
+    }
+    return CommandResult.NoChanges;
   }
 
   private checkUselessClearCell(cmd: ClearCellCommand): CommandResult {
