+++ src/plugins/plugin_registries.ts
@@ -99,8 +99,8 @@ export const corePluginRegistry = new CorePluginRegistry()
   .add("header_grouping", HeaderGroupingPlugin)
   .add("hidden_header", HiddenHeaderPlugin)
   .add("header_visibility", HeaderVisibilityPlugin)
-  .add("cell", CellPlugin)
   .add("default", DefaultPlugin)
+  .add("cell", CellPlugin)
   .add("dataValidation", DataValidationPlugin)
   .add("merge", MergePlugin)
   .add("tables", TablePlugin)
